pyproject.toml
[project]
name = "inmindcnn"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "matplotlib>=3.8.0"
]
